Initial Commit.
This version "works" as a prototype I created with two colleagues.  I
give no guarantee that it will work outside of the downtown Atlanta
Georgia area.  Right now, creating/changing quests requires the end
user to modify a text file on a computer and put it back on the device
in order to play it.